Implementacion de las graficas usando el websocket
