+++ src/shared/components/chart-cards/RotatingTeddyBear.tsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import { View, StyleSheet } from 'react-native';
+import Svg, { Path, Circle, G } from 'react-native-svg';
+import ClinicalColors from '../constants/clinicalcolors';
+
+interface RotatingTeddyBearProps {
+  rotationValue: number;
+  isAnimating: boolean;
+}
+
+const RotatingTeddyBear: React.FC<RotatingTeddyBearProps> = ({ rotationValue, isAnimating }) => {
+  const rotation = Math.abs(rotationValue) * 10;
+  
+  return (
+    <View style={styles.svgContainer}>
+      <Svg 
+        width={60} 
+        height={60} 
+        viewBox="0 0 24 24"
+        style={{ 
+          transform: [{ rotate: `${rotation}deg` }],
+          opacity: isAnimating ? 1 : 0.7
+        }}
+      >
+        <Path 
+          fill={ClinicalColors.secondary}
+          stroke={ClinicalColors.primary}
+          strokeWidth="0.5"
+          d="M15.75 19.13c-.83 0-1.5-.84-1.5-1.88c0-1.03.67-1.87 1.5-1.87s1.5.84 1.5 1.87c0 1.04-.67 1.88-1.5 1.88M12 11.25c-1.24 0-2.25-.84-2.25-1.87c0-1.04 1.01-1.88 2.25-1.88s2.25.84 2.25 1.88c0 1.03-1.01 1.87-2.25 1.87m-3.75 7.88c-.83 0-1.5-.84-1.5-1.88c0-1.03.67-1.87 1.5-1.87s1.5.84 1.5 1.87c0 1.04-.67 1.88-1.5 1.88M12 8.25c.41 0 .75.34.75.75s-.34.75-.75.75s-.75-.34-.75-.75s.34-.75.75-.75M18.75 12c-.32 0-.63.07-.91.2c-.48-.61-1.13-1.13-1.91-1.53c.57-.8.91-1.77.91-2.82v-.06c1.09-.23 1.91-1.2 1.91-2.37c0-1.33-1.09-2.42-2.42-2.42c-.69 0-1.33.29-1.75.75a4.81 4.81 0 0 0-5.16 0C9 3.29 8.36 3 7.67 3C6.34 3 5.25 4.09 5.25 5.42c0 1.16.82 2.13 1.9 2.37v.06c0 1.05.35 2.03.91 2.82c-.77.4-1.42.92-1.9 1.53A2.24 2.24 0 0 0 3 14.25c0 1.25 1 2.25 2.25 2.25h.06c-.04.24-.06.5-.06.75c0 2.07 1.34 3.75 3 3.75c1.01 0 1.9-.63 2.45-1.59c.42.06.85.09 1.3.09s.88-.03 1.3-.09c.55.96 1.44 1.59 2.45 1.59c1.66 0 3-1.68 3-3.75c0-.25-.02-.51-.06-.75h.06c1.25 0 2.25-1 2.25-2.25S20 12 18.75 12"
+        />
+        {isAnimating && (
+          <G>
+            <Circle cx="12" cy="12" r="1" fill={ClinicalColors.warning} opacity="0.6" />
+            <Circle cx="12" cy="12" r="2" fill="none" stroke={ClinicalColors.warning} strokeWidth="0.5" opacity="0.4" />
+          </G>
+        )}
+      </Svg>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  svgContainer: {
+    marginRight: 20,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default RotatingTeddyBear;
