+++ src/shared/components/constants/clinicalcolors.ts
@@ -1,0 +1,14 @@
+const ClinicalColors = {
+    stable: '#4CAF50',     
+    anxious: '#FF9800',     
+    crisis: '#F44336',      
+    primary: '#9BC4E0',     
+    secondary: '#DB7093',   
+    white: '#FFFFFF',
+    textPrimary: '#212121',
+    textSecondary: '#757575',
+    cardBorder: '#E0E0E0',
+    warning: '#FFC107',
+};
+
+export default ClinicalColors;
